Major overhaul: put beam stuff into a module
This tucks all the Neo4j Beam stuff into a specific module (neo4j_beam)
and cleans up the pipeline code a bit. It also vendors the strtobool
function to squash the deprecation warning.

Updated docs and Makefile, too! Now we get a handly link to the pipeline
job page.